+++ tests/test_enhancers/test_blur_detect.py
